#22 feat(techstack) : 기술스택 어드민 전용 권한 체크

+++ src/main/java/com/sejong/projectservice/application/techstack/service/TechStackService.java
@@ -1,12 +1,14 @@
 package com.sejong.projectservice.application.techstack.service;
 
+import com.sejong.projectservice.application.common.error.code.ErrorCode;
+import com.sejong.projectservice.application.common.error.exception.ApiException;
 import com.sejong.projectservice.application.techstack.dto.TechStackCreateReq;
 import com.sejong.projectservice.application.techstack.dto.TechStackRes;
 import com.sejong.projectservice.core.techstack.TechStack;
 import com.sejong.projectservice.core.techstack.TechstackRepository;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -15,19 +17,28 @@ public class TechStackService {
     private final TechstackRepository techstackRepository;
 
     @Transactional
-    public TechStackRes createTechStack(TechStackCreateReq techstackCreateReq) {
+    public TechStackRes createTechStack(TechStackCreateReq techstackCreateReq, String userRole) {
+        validateAdminRole(userRole);
         TechStack techStack = techstackCreateReq.toDomain();
         TechStack savedTechstack = techstackRepository.save(techStack);
         return TechStackRes.from(savedTechstack);
     }
 
+    private static void validateAdminRole(String userRole) {
+        if(!userRole.equals("ADMIN")) {
+            throw new ApiException(ErrorCode.BAD_REQUEST,"어드민 전용 권한입니다.");
+        }
+    }
+
+    @Transactional(readOnly = true)
     public TechStackRes getTechStack(Long techStackId) {
         TechStack techStack = techstackRepository.findById(techStackId);
         return TechStackRes.from(techStack);
     }
 
     @Transactional
-    public TechStackRes updateTechStack(Long techStackId, TechStackCreateReq request) {
+    public TechStackRes updateTechStack(Long techStackId, TechStackCreateReq request, String userRole) {
+        validateAdminRole(userRole);
         TechStack techStack = techstackRepository.findById(techStackId);
         techStack.update(request.getName());
         TechStack savedTechStack = techstackRepository.save(techStack);
@@ -35,7 +46,8 @@ public class TechStackService {
     }
 
     @Transactional
-    public void deleteTechStack(Long techStackId) {
+    public void deleteTechStack(Long techStackId, String userRole) {
+        validateAdminRole(userRole);
         techstackRepository.deleteById(techStackId);
     }
 }
